==> notebooks/00e-GeoNamesCountry.ipynb <==
# Countries

**[Work in progress]**

This notebook creates a .csv file with country information for ingestion into the Knowledge Graph.

Data source: [GeoNames.org](https://download.geonames.org/export/dump/)

Author: Peter Rose (pwrose@ucsd.edu)

import os
from pathlib import Path
import pandas as pd

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

### Create countries

country_url = 'https://download.geonames.org/export/dump/countryInfo.txt'

names = ['ISO','ISO3','ISO-Numeric','fips','Country','Capital','Area(in sq km)','Population',
         'Continent','tld','CurrencyCode','CurrencyName','Phone','Postal Code Format',
         'Postal Code Regex','Languages','geonameid','neighbours','EquivalentFipsCode'
        ]

countries = pd.read_csv(country_url, sep='\t',comment='#', dtype='str', names=names)

### Add missing data
TODO North American Countries and island nations are missing the Continent data

Add missing iso code for Namibia

index = countries.query("ISO3 == 'NAM'").index
countries.at[index, 'ISO'] = 'NA'
countries.head()

### Standardize column names for Knowlege Graph
* id: unique identifier for country
* name: name of node
* parentId: unique identifier for continent
* properties: camelCase

# https://www.iso.org/obp/ui/#iso:code:3166:BQ

countries['id'] = countries['ISO'] # standard id column to link nodes
countries.rename(columns={'ISO': 'iso'}, inplace=True)
countries.rename(columns={'ISO3': 'iso3'}, inplace=True)
countries.rename(columns={'Country': 'name'}, inplace=True)
countries.rename(columns={'Population': 'population'}, inplace=True)
countries.rename(columns={'Area(in sq km)': 'areaSqKm'}, inplace=True)
countries.rename(columns={'Continent': 'parentId'}, inplace=True)

### Export a minimum subset for now

countries = countries[['id','name','iso','iso3','population','areaSqKm','parentId']]
countries = countries.fillna('')

countries.head(1000)

countries.to_csv(NEO4J_HOME / "import/00e-GeoNamesCountry.csv", index=False)
