==> notebooks/00f-GeoNamesAdmin1.ipynb <==
# First Administrative Divisions of Countries

**[Work in progress]**

This notebook creates a .csv file with first administrative divisions (State, Province, Municipality) for ingestion into the Knowledge Graph.

Data source: [GeoNames.org](https://download.geonames.org/export/dump/)

Author: Peter Rose (pwrose@ucsd.edu)

import os
from pathlib import Path
import pandas as pd

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

### Create admin1

admin1_url = 'https://download.geonames.org/export/dump/admin1CodesASCII.txt'

names = ['code', 'name', 'name_ascii', 'geonameid']

admin1 = pd.read_csv(admin1_url, sep='\t', dtype='str', names=names)
admin1 = admin1[['code', 'name_ascii']]

### Standardize column names for Knowlege Graph
* id: unique identifier for country
* name: name of node
* parentId: unique identifier for continent
* properties: camelCase

admin1.rename(columns={'code': 'id'}, inplace=True) # standard id column to link nodes
admin1.rename(columns={'name_ascii': 'name'}, inplace=True)
admin1['code'] = admin1['id'].str.split('.', expand=True)[1]
admin1['parentId'] = admin1['id'].str.split('.', expand=True)[0]

### Example

admin1.query("name == 'California'")

### Export a minimum subset for now

admin1 = admin1[['id','name','code','parentId']]
admin1 = admin1.fillna('')

admin1.head()

admin1.to_csv(NEO4J_HOME / "import/00f-GeoNamesAdmin1.csv", index=False)
