==> notebooks/00j-USCensusCountyCity2017.ipynb <==
# US 2017 State, County, Minor Civil Division, and Incorporated Place FIPS Codes  

**[Work in progress]**

This notebook creates .csv files with US 2017 State and County FIPS codes for ingestion into the Knowledge Graph.

Data source: [2017 State, County, Minor Civil Division, and Incorporated Place FIPS Codes](https://www2.census.gov/programs-surveys/popest/geographies/2017/all-geocodes-v2017.xlsx)

Author: Peter Rose (pwrose@ucsd.edu)

import os
from pathlib import Path
import pandas as pd

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

census_url = 'https://www2.census.gov/programs-surveys/popest/geographies/2017/all-geocodes-v2017.xlsx'

df = pd.read_excel(census_url, dtype='str', skiprows=4)

df.head()

df.rename(columns={'Area Name (including legal/statistical area description)': 'name'}, inplace=True)

### Create list of US Counties

counties = df.query("`County Code (FIPS)` != '000'").query("`County Subdivision Code (FIPS)` == '00000'").copy()
counties.rename(columns={'County Code (FIPS)': 'fips'}, inplace=True)
counties['stateFips'] = counties['State Code (FIPS)']
counties = counties[['name', 'fips', 'stateFips']]

counties.count()

counties.head()

counties.query("fips == '039'").query("stateFips == '01'")

counties.to_csv(NEO4J_HOME / "import/00j-USCensus2017County.csv", index=False)

### Create list of US Cities

cities = df.query("`Place Code (FIPS)` != '00000'").copy()
cities.rename(columns={'Place Code (FIPS)': 'fips'}, inplace=True)
cities['stateFips'] = cities['State Code (FIPS)']
cities = cities[['name', 'fips', 'stateFips']]

cities.count()

cities.head()

cities['name'] = cities['name'].str.replace(' city','')
cities['name'] = cities['name'].str.replace(' town','')
cities.head()

cities.query("name == 'Covington'")

cities.to_csv(NEO4J_HOME / "import/00j-USCensus2017City.csv", index=False)
