==> notebooks/00k-UNRegion.ipynb <==
# UN Regions

**[Work in progress]**

This notebook creates a .csv file with UN geographic regions, subregions, and intermediate region information for ingestion into the Knowledge Graph.

Data source: [Statistics Division of the United Nations Secretariat](https://unstats.un.org/unsd/methodology/m49/)

Data set: [M49](https://unstats.un.org/unsd/methodology/m49/overview)

Authors: Braden Riggs (bdriggs@ucsd.edu), Peter Rose (pwrose@ucsd.edu)

import os
from pathlib import Path
import pandas as pd

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

### UN regions, subregions, and intermediate regions

df = pd.read_excel("../reference_data/UNSDMethodology.xlsx", dtype='str')
df = df[['Region Name', 'Region Code', 'Sub-region Name', 'Sub-region Code', 
         'Intermediate Region Name', 'Intermediate Region Code', 'ISO-alpha3 Code']]
df.fillna('', inplace=True)
# for now exclude region without an iso code (Channel Islands)
df = df.query("`ISO-alpha3 Code` != ''")
# Antarctica has no region code
df = df.query("`Region Name` != ''")
df.head()

Assign names without spaces

df.rename(columns={'Region Name': 'UNRegion'}, inplace=True)
df.rename(columns={'Region Code': 'UNRegionCode'}, inplace=True)
df.rename(columns={'Sub-region Name': 'UNSubRegion'}, inplace=True)
df.rename(columns={'Sub-region Code': 'UNSubRegionCode'}, inplace=True)
df.rename(columns={'Intermediate Region Name': 'UNIntermediateRegion'}, inplace=True)
df.rename(columns={'Intermediate Region Code': 'UNIntermediateRegionCode'}, inplace=True)
df.rename(columns={'ISO-alpha3 Code': 'iso3'}, inplace=True)

### Assign unique identifiers
Use m49 as a prefix for the M49 standard  by the Statistics Division of the United Nations Secretariat 

df['UNRegionCode'] = 'm49:' + df['UNRegionCode']
df['UNSubRegionCode'] = 'm49:' + df['UNSubRegionCode'] 
df['UNIntermediateRegionCode'] = 'm49:' + df['UNIntermediateRegionCode'] 

df.head()

### Add missing region information (from hand-curated list)

additions = pd.read_csv("../reference_data/UNRegionAdditions.csv")
additions.fillna('', inplace=True)

additions.tail(10)

df = df.append(additions)

df.to_csv(NEO4J_HOME / "import/00k-UNAll.csv", index=False)  

### Save region assignments in separate files
This is done so iso3 country codes can be linked to the lowest level in the UN region hierarchy.

intermediateRegion = df[df['UNIntermediateRegion'] != '']
intermediateRegion.to_csv(NEO4J_HOME / "import/00k-UNIntermediateRegion.csv", index=False)                            

intermediateRegion.head()

subRegion = df[(df['UNSubRegion'] != '') & (df['UNIntermediateRegion'] == '')]
subRegion.to_csv(NEO4J_HOME / "import/00k-UNSubRegion.csv", index=False)  

subRegion.head()

region = df[(df['UNSubRegion'] == '') & (df['UNIntermediateRegion'] == '')]
region.to_csv(NEO4J_HOME / "import/00k-UNRegion.csv", index=False)  

region.head()
