==> notebooks/01c-CNCBStrainLocations.ipynb <==
# Assigns location to Strain Data from CNCB
**[Work in progress]**

This notebook standardizes location information for viral strain from CNCB for ingestion into a Knowledge Graph.

Author: Peter Rose (pwrose@ucsd.edu)

import os
import pandas as pd
#import dateutil
from pathlib import Path
#from os import path

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

## Standardize Location data for SARS-CoV-2 Strain metadata

TODO this code should be replaced with a general solution.

Below is a workaround for now.

df = pd.read_csv(NEO4J_HOME / "import/01c-CNCBStrain.csv", dtype='str')
df.fillna('', inplace=True)

df.head()

#### Split locations

df['locationLevels'] = df['location'].str.count('/')

df[['country', 'admin1', 'admin2', 'city']] = df['location'].str.split('/', n=3, expand=True)
df['country'] = df['country'].str.strip()
df['admin1'] = df['admin1'].str.strip()
df['admin2'] = df['admin2'].str.strip()
df['city'] = df['city'].str.strip()

Fix country names

df['country'] = df['country'].str.replace('\u200eRomania','Romania')
df['country'] = df['country'].str.replace('Czech Republic','Czechia')
df['country'] = df['country'].str.replace('Czech republic','Czechia')
df['country'] = df['country'].str.replace('Luxemburg','Luxembourg')
df.fillna('', inplace=True)

Check for country mismatches

countries = pd.read_csv(NEO4J_HOME / "import/00e-GeoNamesCountry.csv")
loc0 = df[['country']].copy()
loc0 = loc0.merge(countries, left_on='country', right_on='name', how='left')

loc0.fillna('', inplace=True)
loc0 = loc0.query("name == ''")
print("Country name mismatches:")
loc0['country'].unique()

Fix Admin1 names

states = {
        # US
        'AK': 'Alaska',
        'AL': 'Alabama',
        'AR': 'Arkansas',
        'AS': 'American Samoa',
        'AZ': 'Arizona',
        'CA': 'California',
        'CO': 'Colorado',
        'CT': 'Connecticut',
        'DC': 'District of Columbia',
        'DE': 'Delaware',
        'FL': 'Florida',
        'GA': 'Georgia',
        'GU': 'Guam',
        'HI': 'Hawaii',
        'IA': 'Iowa',
        'ID': 'Idaho',
        'IL': 'Illinois',
        'IN': 'Indiana',
        'KS': 'Kansas',
        'KY': 'Kentucky',
        'LA': 'Louisiana',
        'MA': 'Massachusetts',
        'MD': 'Maryland',
        'ME': 'Maine',
        'MI': 'Michigan',
        'MN': 'Minnesota',
        'MO': 'Missouri',
        'MP': 'Northern Mariana Islands',
        'MS': 'Mississippi',
        'MT': 'Montana',
        'NA': 'National',
        'NC': 'North Carolina',
        'ND': 'North Dakota',
        'NE': 'Nebraska',
        'NH': 'New Hampshire',
        'NJ': 'New Jersey',
        'NM': 'New Mexico',
        'NV': 'Nevada',
        'NY': 'New York',
        'OH': 'Ohio',
        'OK': 'Oklahoma',
        'OR': 'Oregon',
        'PA': 'Pennsylvania',
        'PR': 'Puerto Rico',
        'RI': 'Rhode Island',
        'SC': 'South Carolina',
        'SD': 'South Dakota',
        'TN': 'Tennessee',
        'TX': 'Texas',
        'UT': 'Utah',
        'VA': 'Virginia',
        'VI': 'Virgin Islands',
        'VT': 'Vermont',
        'WA': 'Washington',
        'WI': 'Wisconsin',
        'WV': 'West Virginia',
        'WY': 'Wyoming',
        # Canada
        'AB': 'Alberta',
        'BC': 'British Columbia',
        'MB': 'Manitoba',
        'NB': 'New Brunswick',
        'NL': 'Newfoundland and Labrador',
        'NT': 'Northwest Territories',
        'NS': 'Nova Scotia',
        'NU': 'Nunavut',
        'ON': 'Ontario',
        'PE': 'Prince Edward Island',
        'QC': 'Quebec',
        'SK': 'Saskatchewan',
        'YT': 'Yukon'
}

# # expand states to full name
df['admin1'] = df['admin1'].apply(lambda s: states.get(s, s))
df['admin1'] = df['admin1'].str.replace('Nonthaburi Province','Nonthaburi')
df['admin1'] = df['admin1'].str.replace('Guangdong Province','Guangdong')
df['admin1'] = df['admin1'].str.replace('Ile De France','Ile-de-France')
df.fillna('', inplace=True)

Check for Admin1 mismatches

admin1 = pd.read_csv(NEO4J_HOME / "import/00f-GeoNamesAdmin1.csv")
loc1 = df[['admin1']].copy()
loc1 = loc1.query("admin1 != ''")
loc1 = loc1.merge(admin1, left_on='admin1', right_on='name', how='left')

loc1.fillna('', inplace=True)
loc1 = loc1.query("name == ''")
print("Admin1 name mismatches:")
loc1['admin1'].unique()

Check for Admin2 names

admin2 = pd.read_csv(NEO4J_HOME / "import/00g-GeoNamesAdmin2.csv")
loc2 = df[['admin2']].copy()
loc2 = loc2.query("admin2 != ''")
loc2 = loc2.merge(admin2, left_on='admin2', right_on='name', how='left')

loc2.fillna('', inplace=True)
loc2 = loc2.query("name == ''")
print("Admin2 name mismatches:")
loc2['admin2'].unique()

Check city names

city = pd.read_csv(NEO4J_HOME / "import/00h-GeoNamesCity.csv")
loc3 = df[['city']].copy()

loc3 = loc3.query("city != ''")

loc3

loc3 = loc3.query("city != ''")
loc3 = loc3.merge(city, left_on='city', right_on='name', how='left')

loc3.fillna('', inplace=True)
loc3 = loc3.query("name == ''")
print("City name mismatches:")
loc3['city'].unique()

strains = df[['id', 'name', 'alias', 'taxonomyId', 'hostTaxonomyId', 'collectionDate', 
              'country', 'admin1', 'admin2', 'city','locationLevels']]
strains.head()

strains.to_csv(NEO4J_HOME / "import/01c-CNCBStrain.csv", index=False)
