==> notebooks/01d-CNCBStrain.ipynb <==
# Load SARS-CoV-2 Virus Strain Data from CNCB
**[Work in progress]**

This notebook downloads and standardizes viral strain and variation data from CNCB for ingestion into a Knowledge Graph.

Data source: [China National Center for Bioinformation, 2019 Novel Coronavirus Resource (2019nCoVR)](https://bigd.big.ac.cn/ncov/release_genome)

Author: Peter Rose (pwrose@ucsd.edu)

import os
import pandas as pd
import dateutil
import re
from pathlib import Path
import glob
#from os import path
import ftplib

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

# Path will take care of handling operating system differences.
NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

# Create a directory to cache variation data
CACHE = Path(NEO4J_HOME / 'import/cache')
CACHE.mkdir(exist_ok=True)

## Download SARS-CoV-2 Strain metadata

This notebook will download > 17,000 files. To work on a small sample (50 files), set 

`run_small_sample_only = True`

run_small_sample_only = False

metadata_url = "https://bigd.big.ac.cn/ncov/genome/export/meta"
annotation_url = "ftp://download.big.ac.cn/GVM/Coronavirus/gff3/" 

df = pd.read_excel(metadata_url, dtype='str')
df.fillna('', inplace=True)

print("Total number of strains:", df.shape[0])

df = df.query("`Sequence Quality` == 'High'")
df = df.query("`Nuc.Completeness` == 'Complete'")

print("Number of complete high quality strains", df.shape[0])

df.head()

#### Create a separate row for each Accession and Related ID

df['Accession ID'] = df['Accession ID'].str.strip()
df['Related ID'] = df['Related ID'].str.strip()

# combine all ids into a single column
df['alias'] = df['Accession ID'] + df['Related ID'].apply(lambda s: ',' + s if len(s) > 0 else s)
df['alias'] = df['alias'].str.replace(' ', '')

# then "explode" ids into separate rows
df['id'] = df['alias'].apply(lambda s: s.split(','))
df = df.explode('id')
df['id'] = df['id'].str.strip()
df['alias'] = df['alias'].str.replace(',', ';')

#### Assign taxonomy ids

# TODO a find general solution to map host name to NCBI taxonomy**

# some organism specifications are ambiguous, 
# they don't match a specificn NCBI taxonomy

taxonomy_to_id = {'Human': 'taxonomy:9606', 
                  'Homo sapiens': 'taxonomy:9606',
                  'Homo Sapiens': 'taxonomy:9606',
                  'Rhinolophus affinis': 'taxonomy:59477', 
                  'Rhinolophus malayanus': 'taxonomy:608659', 
                  'Mustela lutreola': 'taxonomy:9666',
                  'Panthera tigris jacksoni': 'taxonomy:419130',
                  'Rhinolophus sp. (bat)': 'taxonomy:49442', # ambiguous
                  'bat': 'taxonomy:49442', # ambiguous
                  'Manis javanica': 'taxonomy:9974',
                  'palm civet': 'taxonomy:71116', # ambiguous
                  'Canine': 'taxonomy:9608', # ambiguous
                  'canine': 'taxonomy:9608', # ambiguous
                  'Felis catus': 'taxonomy:9685'
                 }

# assign taxonomy id to host
df['hostTaxonomyId'] = df['Host'].apply(lambda s: taxonomy_to_id.get(s, s))
df['hostTaxonomyId'].unique()

df['taxonomyId'] = 'taxonomy:2697049' # SARS-CoV-2

#### Standardize node property names (CURIEs and URIs)

df.rename(columns={'Virus Strain Name': 'name',
                   'Sample Collection Date':'collectionDate',
                   'Location':'location'}, 
          inplace=True)

# https://registry.identifiers.org/registry/insdc
insdc_pattern = re.compile('^([A-Z]\d{5}|[A-Z]{2}\d{6}|[A-Z]{4}\d{8}|[A-J][A-Z]{2}\d{5})(\.\d+)?$')

def assign_curie(id):
    id = id.strip()
    # remove underscore to enable CURIE matching of NCBI reference sequences NC_...
    id = id.replace('NC_', 'NC') 
    if len(id) > 0:
        if id.startswith('EPI'):
            return 'https://www.gisaid.org/' + id
        elif id.startswith('NC_'):
            # NCBI reference sequences resolve with ncbiprotein CURIE
            return 'ncbiprotein:' + id
        elif insdc_pattern.match(id) != None:
            return 'insdc:' + id
        else:
            # TODO are URIs available for these cases?
            return id
    else:
        return id

strains = df[['id', 'name', 'alias', 'taxonomyId', 'hostTaxonomyId','collectionDate', 'location']].copy()
strains['id'] = strains['id'].apply(assign_curie)
strains.head()

strains.to_csv(NEO4J_HOME / "import/01d-CNCBStrain.csv", index=False)

## Merge Metadata with Variation Data

#### Get list of file names from FTP site

server = "download.big.ac.cn"
user = "anonymous"
password = "anonymous"
source = "/GVM/Coronavirus/gff3/"

ftp = ftplib.FTP(server)
ftp.login(user, password)
ftp.cwd(source) 
filelist=ftp.nlst()
ftp.quit()

df_file = pd.DataFrame(filelist, columns=['filename'])

Extract identifiers from file name

Example: 2019-nCoV_CNA0013697_variants.gff3 -> CNA0013697

df_file['id'] = df_file['filename'].str[10:]
df_file['id'] = df_file['id'].str.replace('_variants.gff3','')

df_file = df_file.query("id != ''")

print("Number of available files:", df_file.shape[0])

df_file.head()

df = df.merge(df_file, on='id')

print('Strains with a matching filename:', df.shape[0])

if run_small_sample_only:
    df = df.sample(n=50, random_state=5)

df.head()

Keep only unique entries (there are a few duplicate cases)

df.drop_duplicates(subset='name', inplace=True)
df.drop_duplicates(subset='id', inplace=True)

print('Strains with a matching filename:', df.shape[0])

#### Download variant annotations for each strain
To avoid download the same files every time, they are cached, and newly downloaded files are added to the cache.

names = ['taxon1', 'variantType', 'name', 'start', 'end','x1', 'x2', 'x3','taxon2', 'x4', 'strainStart', 'taxon3', 'x5', 'strainEnd', 'ref', 'alt', 'vepAnnotation']

def download_gff3(filename, url):
    gff3 = pd.read_csv(url, header=None, comment='#', sep='[\t|;]', engine='python', names=names)
    gff3['ref'] = gff3['ref'].str.replace('REF=','')
    gff3['alt'] = gff3['alt'].str.replace('ALT=','')
    gff3['vepAnnotation'] = gff3['vepAnnotation'].str.replace('VEP=','')
    # prepare for 3-way split (need at least two commas)
    gff3['vepAnnotation'] = gff3['vepAnnotation'].apply(lambda s: s + ',,' if s.count(',') < 2 else s)
    # 3-way split
    gff3[['variantConsequence','proteinVariant','geneVariant']] = gff3['vepAnnotation'].str.split(',', n=2, expand=True)
    gff3['geneVariant'] = gff3['geneVariant'].str.replace('gene-','')
    gff3 = gff3[['name', 'variantType', 'start', 'end', 'ref', 'alt', 'variantConsequence', 'proteinVariant', 'geneVariant']]
    
    filename = row['filename'] + '.csv'
    gff3.to_csv(CACHE / filename, index=False)

for index, row in df.iterrows():
    url = annotation_url + row['filename']
    filename = row['filename'] + '.csv'
    if not Path.exists(CACHE / filename):
        try:
            download_gff3(row['filename'], url)
            print(row['filename'], end=' ')
        except:
            print('Download failed for: ', row['filename'])

### Concatenate all variation data into a single dataframe

# use all cached data files
path = str(CACHE / '*.csv')
filenames = glob.glob(path)

# list of dataframes
dfs = []

for filename in filenames:
    df = pd.read_csv(filename, index_col=None, header=0)
    dfs.append(df)

variations = pd.concat(dfs, axis=0, ignore_index=True)
variations.fillna('', inplace=True)

List of variant types and consequences:

https://uswest.ensembl.org/info/genome/variation/prediction/classification.html

https://uswest.ensembl.org/info/genome/variation/prediction/predicted_data.html#consequences

#### Extract protein position and protein id from proteinVariant string

Example: QHD43415.1:p.5828P>L

proteinPosition: 5828
proteinId: QHD43415

position_pattern = re.compile(':p\.(.*?)[A-Z|\-]+')

def extract_protein_position(s):
    if s == '':
        return s
    else:
        groups = position_pattern.search(s)
        if groups == None:
            return ''
        else:
            return groups.group(1)

variations['proteinPosition'] = variations['proteinVariant'].apply(extract_protein_position)

variations['proteinId'] = variations['proteinVariant'].apply(lambda s: s.split('.')[0] if '.' in s else '')

variations['proteinId'].unique()

#### Assign SARS-CoV-2 taxonomy id

variations['taxonomyId'] = 'taxonomy:2697049'

#### Assign Reference genome

The first SARS-CoV-2 genome sequence is the reference for the variant annotation below.

[Severe acute respiratory syndrome coronavirus 2 isolate Wuhan-Hu-1](https://www.ncbi.nlm.nih.gov/nuccore/MN908947)

variations['referenceGenome'] = 'insdc:MN908947'

variations['proteinId'] = variations['proteinId'].apply(lambda s: 'uniprot:' + s if s != '' else s)

print('variantType:', variations['variantType'].unique())

print("variantConsequence:", variations['variantConsequence'].unique())

print("Number of variants:", variations.shape[0])

variations.to_csv(NEO4J_HOME / "import/01d-CNCBVariant.csv", index=False)

variations.head()
