==> notebooks/01e-ProteinProteinInteraction.ipynb <==
# Protein Protein Interaction Data
**[Work in progress]**

This notebook downloads and standardizes viral-host protein data from IntAct for ingestion into the Knowledge Graph.

Data sources: [IntAct](https://www.ebi.ac.uk/intact/query/pubid:IM-27814)

Authors: Kaushik Ganapathy, Eric Yu, Peter Rose (krganapa@ucsd.edu, ery010@ucsd.edu, pwrose@ucsd.edu)

import os
import re
import hashlib 

import pandas as pd
import numpy as np

from pathlib import Path
from Bio import SeqIO

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

https://www.uniprot.org/help/uniprotkb_column_names
https://www.uniprot.org/uniprot/P0DTD1#PRO_0000449630

### Retrieve interaction data from IntAct

##### Query for interactions with SARS-CoV-2: "taxid:2697049"

data = pd.read_csv("https://www.ebi.ac.uk/intact/export?format=mitab_25&query=taxid%3A2697049&negative=false&spoke=false&ontology=false&sort=intact-miscore&asc=false", sep='\t')

print('Number of interactions:', data.shape[0])

data.head()

### Process Data

data.rename(columns={'#ID(s) interactor A': 'interactorA', 'ID(s) interactor B': 'interactorB'}, inplace=True)

data = data.replace('uniprotkb','uniprot', regex=True)

#### Extract UniProt accession number and Uniprot protein id from interactor columns

# uniprot:P0DTD1-PRO_0000449619 -> P0DTD1-PRO_0000449619
data['id_a'] = data['interactorA'].str.replace('uniprot:', '')
data['id_b'] = data['interactorB'].str.replace('uniprot:', '')

# P0DTD1-PRO_0000449619 -> P0DTD1 (UniProt accession number)
data['accession_a'] = data['id_a'].str.split('-', expand=True)[0]
data['accession_b'] = data['id_b'].str.split('-', expand=True)[0]

# PRO_0000449619 -> PRO_0000449619 (UniProt protein id)
data['pro_id_a'] = data['id_a'].str.split('-', expand=True)[1]
data['pro_id_b'] = data['id_b'].str.split('-', expand=True)[1]

data.fillna('', inplace=True)

#### Extract pubmed id
Example: imex:IM-27912|pubmed:32275855 -> 2275855

position_pattern = re.compile('pubmed:(\d*).')

def extract_pubmed_id(s):
    groups = position_pattern.search(s)
    if groups == None:
        return ''
    else:
        return groups.group(1)

data['pubmedId'] = data['Publication Identifier(s)'].apply(extract_pubmed_id)

#### Extract taxonomy id
Example: taxid:9606(human)|taxid:9606(Homo sapiens) -> 9606

position_pattern = re.compile('taxid:(\d*)\(')

def extract_tax_id(s):
    groups = position_pattern.search(s)
    if groups == None:
        return ''
    else:
        return groups.group(1)

data['taxonomy_id_a'] = data['Taxid interactor A'].apply(extract_tax_id)
data['taxonomy_id_b'] = data['Taxid interactor B'].apply(extract_tax_id)

data = data[['id_a', 'id_b', 'accession_a', 'accession_b', 'pro_id_a', 'pro_id_b', 'taxonomy_id_a', 'taxonomy_id_b', 'pubmedId']]

data.head()

#### Restrict data to SARS-CoV-2 - human protein-protein interactions (TODO: expand to SARS, MERS, etc.)

data = data[((data['taxonomy_id_a'] == '2697049') & (data['taxonomy_id_b'] == '9606')) |
            ((data['taxonomy_id_b'] == '2697049') & (data['taxonomy_id_a'] == '9606'))]

data.drop_duplicates(inplace=True)

#### Retrieve Protein names and sequences from UniProt.org

Collect unique accession numbers

a = data['accession_a'].unique()
b = data['accession_b'].unique()
ids = np.unique(np.concatenate((a, b)))

def split_chains(accession, chains):
    chain_list = []
    for chain in chains:
        items = chain.split(';')
        feature_dict = {}
        
        try:
            if len(items) > 1:
                feature_dict['accession'] = accession
                for item in items:
                    item = item.strip()
                    if '..' in item:
                        feature_dict['start'] = int(items[0].split('..')[0])
                        feature_dict['end'] = int(items[0].split('..')[1])
                    elif item.startswith("/note="):
                        name = item[6:].replace('\"', '')
                        feature_dict['name'] = name
                    elif item.startswith("/id="):
                        pro_id = item[4:].replace('\"', '')
                        feature_dict['pro_id'] = pro_id
                chain_list.append(feature_dict)
        except:
            print('Skipping incomplete features for', accession, ':', chain)
            
    df = pd.DataFrame(chain_list)
    df.dropna(inplace=True)
    return df

seq_dfs = []
for accession in ids:
    url = f'https://www.uniprot.org/uniprot/?query=accession:{accession}&columns=id,sequence,length,protein%20names,feature(CHAIN),organism-id&format=tab'
    try:
        df = pd.read_csv(url, sep='\t')
        df.fillna('', inplace=True)
        
        whole_chain = df.copy()
        
        start = 1
        end = whole_chain.iloc[0]['Length']

        whole_chain['accession'] = accession
        whole_chain['start'] = start
        whole_chain['end'] = end
        whole_chain['name'] = whole_chain['Protein names']
        whole_chain['name'] = whole_chain['Protein names'].str.split('\(', expand=True)
        whole_chain['pro_id'] =''
        whole_chain = whole_chain[['accession', 'start', 'end', 'name', 'pro_id']]
        
        # process features (Chains)
        feature_string = df['Chain'].values[0]
        if (feature_string != ''):
            chains = feature_string.split("CHAIN ")
            chain_df = split_chains(accession, chains)
            if chain_df.shape[0] > 0:
                # Add the whole chain if it is not in the list of chains
                if chain_df.iloc[0]['start'] != start or chain_df.iloc[0]['end'] != end:
                    chain_df = pd.concat([whole_chain, chain_df])
                    
                df = df.merge(chain_df, left_on='Entry', right_on='accession', how='left')
                print(accession, end=' ')
                seq_dfs.append(df)
        else:
            df = df.merge(whole_chain, left_on='Entry', right_on='accession', how='left')
            seq_dfs.append(df)
    except:
        print("Uniprot accession not found:", accession)
        
seq = pd.concat(seq_dfs, axis=0, ignore_index=True)

#### Cleave sequences into peptides

def get_subsequence(row):
    start = row['start']
    end = row['end']
    sequence = row['Sequence']
    return sequence[start-1: end]

seq['sequence'] = seq.apply(lambda row: get_subsequence(row), axis=1)

Set flag if protein chain is full length

seq['fullLength'] = (seq['start'] == 1) & (seq['end'] == seq['Length'])

seq['id'] = seq['sequence'].apply(lambda seq: 'md5:' + hashlib.md5(seq.encode()).hexdigest())

seq.rename(columns={'Organism ID': 'taxonomyId'}, inplace=True)

seq['name'] = seq['name'].str.strip()

seq.head()

### Save proteins

proteins = seq[['id', 'name', 'accession', 'pro_id', 'sequence', 'start', 'end', 'fullLength', 'taxonomyId']].copy()
proteins['accession'] = 'uniprot:' + proteins['accession']
proteins['taxonomyId'] = 'taxonomy:' + proteins['taxonomyId'].astype(str)
proteins.to_csv(NEO4J_HOME / 'import/01e-ProteinProteinInteractionProtein.csv', index = False)

print('Number of proteins:', proteins.shape[0])

proteins.head()

### Merge interaction data with sequence data

sequences = seq[['accession', 'pro_id', 'id']].copy()
sequences_full_length = seq.query('fullLength')
sequences_full_length = sequences_full_length[['accession', 'pro_id', 'id']].copy()

sequences_full_length.head()

sequences_full_length.query("accession == 'Q9BYF1'")

interact = data[['accession_a', 'accession_b', 'pro_id_a', 'pro_id_b', 'pubmedId']]

interact.query("accession_a == 'P0DTC2'")

Split dataframe into two parts, depending on the presence of a protein id for interactor A. Each part is processed differently.

interact_pro_a = interact.query("pro_id_a != ''").copy()
interact_acc_a = interact.query("pro_id_a == ''").copy()

interact_pro_a = interact_pro_a.merge(sequences, left_on='pro_id_a', right_on='pro_id', how='left')
interact_pro_a.fillna('', inplace=True)

interact_acc_a = interact_acc_a.merge(sequences_full_length, left_on='accession_a', right_on='accession', how='left')
interact_acc_a.fillna('', inplace=True)

Concatenate the dataframes back together

interact = pd.concat([interact_pro_a, interact_acc_a])
interact.rename(columns={'id': 'id_a'}, inplace=True)

print('total interactions:', interact.shape[0])

Split dataframe into two parts, depending on the presence of a protein id for interactor B. Each part is processed differently.

interact_pro_b = interact.query("pro_id_b != ''").copy()
interact_acc_b = interact.query("pro_id_b == ''").copy()

interact_pro_b = interact_pro_b.merge(sequences, left_on='pro_id_b', right_on='pro_id', how='left')
interact_pro_b.fillna('', inplace=True)

interact_acc_b = interact_acc_b.merge(sequences_full_length, left_on='accession_b', right_on='accession', how='left')
interact_acc_b.fillna('', inplace=True)

Concatenate the dataframes back together

interact = pd.concat([interact_pro_b, interact_acc_b])
interact.rename(columns={'id': 'id_b'}, inplace=True)
interact.head()

interact.query("accession_a == 'P0DTC1'")

### Save interactions

Remove interactions without md5 ids

interact = interact[(interact['id_a'] != '') & (interact['id_b'] != '')]

interact = interact[['id_a', 'id_b']]
interact.drop_duplicates(inplace=True)
interact.to_csv(NEO4J_HOME / 'import/01e-ProteinProteinInteraction.csv', index = False)

print('Number of interactions:', interact.shape[0])

interact.head()
