==> notebooks/01h-PMCAccession.ipynb <==
# Load Accession Numbers Mappings
**[Work in progress]**

This notebook downloads and standardizes accession numbers from life science and biological databases textmined from PubMedCentral full text articles by [Europe PMC](https://europepmc.org/) for ingestion into a Knowledge Graph.

Data source: [ftp site](ftp://ftp.ebi.ac.uk/pub/databases/pmc/TextMinedTerms/)

Author: Peter Rose (pwrose@ucsd.edu)

import os
import pandas as pd
import dateutil
from pathlib import Path

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

ftp = 'ftp://ftp.ebi.ac.uk/pub/databases/pmc/TextMinedTerms/'

#### Collect datasets with epi references

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

ref1 = pd.read_csv(NEO4J_HOME / "import/01b-Nextstrain.csv")
ref1 = ref1[['id']]
ref1 = ref1.drop_duplicates()
ref1.rename(columns={'id': 'accession'}, inplace=True)
ref1.head()

ref2 = pd.read_csv(NEO4J_HOME / "import/01d-CNCBStrain.csv")
ref2.head()

ref2 = ref2[['id']]
ref2 = ref2.drop_duplicates()
ref2.rename(columns={'id': 'accession'}, inplace=True)
ref2.head()

ref3 = pd.read_csv(NEO4J_HOME / "import/01c-NCBIRefSeq.csv")
ref3 = ref3[['strainId']]
ref3 = ref3.drop_duplicates()
ref3.rename(columns={'strainId': 'accession'}, inplace=True)
ref3.head()

ref4 = pd.read_csv(NEO4J_HOME / "import/01e-ProteinProteinInteractionProtein.csv")
ref4 = ref4[['accession']]
ref4 = ref4.drop_duplicates()
ref4.head()               

ref = pd.concat([ref1, ref2, ref3, ref4])

## Assign unique identifiers for interoperabilitiy
A [CURIE](https://en.wikipedia.org/wiki/CURIE) (Compact URI) is a compact abbreviation for Uniform Resource Identifiers (URIs). CURIEs consist of registered prefix and an accession number (prefix:accession). They provide a name space for identifiers to enable uniqueness of identifiers and interoperability among data resources.

[Identifiers.org](http://identifiers.org/) provides a registry and resolution service for life science CURIEs. 

### NCBI Reference Sequences
**id**: CURIE: [pmc](https://registry.identifiers.org/registry/pmc) (PubMed Central, PMC)

**accession**: CURIE: [ncbiprotein](https://registry.identifiers.org/registry/ncbiprotein) (NCBI Reference Sequences, Refseq)

df1 = pd.read_csv(ftp + "refseq.csv", dtype=str)
df1.head()

# Remove version number from refseq to match to the latest version
df1['id'] = 'pmc:' + df1['PMCID']
df1['accession'] = 'ncbiprotein:' + df1['refseq'].str.split('.', expand=True)[0]
df1 = df1[['id','accession']]

df1 = df1.merge(ref, on="accession")
df1 = df1[['id','accession']]
print("Number of refseq matches:", df1.shape[0])
df1.head()

### GISAID Genome Sequences
**id**: CURIE: [pmc](https://registry.identifiers.org/registry/pmc) (PubMed Central, PMC)

**accession**: URI: [https://www.gisaid.org/](https://www.gisaid.org/help/publish-with-gisaid-references) (Global Initiative on Sharing All Influenza Data, GISAID)


df2 = pd.read_csv(ftp + "gisaid.csv", dtype=str)

df2['id'] = 'pmc:' + df2['PMCID']
df2['accession'] = 'https://www.gisaid.org/' + df2['gisaid']
df2 = df2[['id','accession']]
df2.head()

df2 = df2.merge(ref, on="accession")
df2.dropna(inplace=True)
print("Number of GISAID matches:", df2.shape[0])
df2.head()

### UniProt
**id**: CURIE: [pmc](https://registry.identifiers.org/registry/pmc) (PubMed Central, PMC)

**accession**: CURIE: [uniprot](https://registry.identifiers.org/registry/uniprot) ( UniProt Knowledgebase, UniProtKB)

df3 = pd.read_csv(ftp + "uniprot.csv", dtype=str)

df3['id'] = 'pmc:' + df3['PMCID']
df3['accession'] = 'uniprot:' + df3['uniprot']
df3 = df3[['id','accession']]
df3.head()

df3 = df3.merge(ref, on="accession")
df3.dropna(inplace=True)
print("Number of UniProt matches:", df3.shape[0])
df3.head()

### Protein Data Bank (NOT USED YET)
**id**: CURIE: [pmc](https://registry.identifiers.org/registry/pmc) (PubMed Central, PMC)

**accession**: CURIE: [pdb](https://registry.identifiers.org/registry/pdb) (Protein Data Bank, PDB)

# df4 = pd.read_csv(ftp + "pdb.csv", dtype=str)

# df4['id'] = 'pmc:' + df4['PMCID']
# df4['accession'] = 'pdb:' + df4['pdb']
# df4 = df4[['id','accession']]
# df4.head()

### Digital Object Identifier (DOI) (NOT USED YET)
**id**: CURIE: [pmc](https://registry.identifiers.org/registry/pmc) (PubMed Central, PMC)

**accession**: CURIE: [doi](https://registry.identifiers.org/registry/doi) (Digital Object Identifier System, DOI)

#df5 = pd.read_csv(ftp + "doi.csv", dtype=str)

# df5['id'] = 'pmc:' + df5['PMCID']
# df5['accession'] = 'doi' + df5['doi']
# df5 = df5[['id','accession']]
# df5.head()

### Save data for Knowledge Graph Import

df = pd.concat([df1, df2, df3])
df = df.fillna('')
df = df.query("id != ''")
df = df.query("accession != ''")
print('Mappings:', df.count())

df.head()

df.to_csv(NEO4J_HOME / "import/01h-PMC-Accession.csv", index=False)
