==> notebooks/2-CreateKnowledgeGraph.ipynb <==
# Create Neo4j Knowledge Graph
This notebook creates the COVID-19-Net knowledge graph by ingesting the .csv files in the Neo4j import directory.

import os
import time
from pathlib import Path

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

### Start Neo4j database if it is not running

status = !"$NEO4J_HOME"/bin/neo4j status
status = str(status)
if not 'Neo4j is running' in status:
   !"$NEO4J_HOME"/bin/neo4j start

#### Wait until Neo4j has started up

status = !"$NEO4J_HOME"/bin/neo4j status
status = str(status)
while not 'Neo4j is running' in status:
    time.sleep(15)
    status = !"$NEO4J_HOME"/bin/neo4j status
    status = str(status)
    print(status)

# wait until neo4j is ready to receive requests
time.sleep(60)
print(status)

### Run Cypher commands to create Knowledge Graph

start = time.time()

# TODO create a batch script for Windows
#!./run_cyphers.sh
!../scripts/run_cyphers.sh

end = time.time()
print('time:', end-start)

### Make sure to shutdown Neo4j!

!"$NEO4J_HOME"/bin/neo4j stop
