==> notebooks/02a-JHUCasesLocation.ipynb <==
# COVID-19 Case Counts from Johns Hopkins University

**[Work in progress]**

This notebook adds location information for case counts for ingestion into the Knowledge Graph.

Data source: [COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University](https://github.com/CSSEGISandData/COVID-19)

Author: Peter Rose (pwrose@ucsd.edu)

import os
from pathlib import Path
import pandas as pd
import dateutil

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

NEO4J_HOME = Path(os.getenv('NEO4J_HOME'))
print(NEO4J_HOME)

df = pd.read_csv(NEO4J_HOME / "import/02a-JHUCasesGlobal.csv", dtype='str')
df.fillna('', inplace=True)
df['country'] = df['country'].str.strip()
df['admin1'] = df['admin1'].str.strip()

df.head()

### Standardize country names to match GeoNames.org

ref = pd.read_csv("../reference_data/SpecialLocations.csv", comment='#', dtype='str')

Convert dataframe into a dictionary

ref['val'] = ref[['geoname', 'type']].values.tolist()
name_list = ref[['name', 'val']].values.tolist()
name_dict = {name: val for name, val in name_list}

def standardize_locations(row):
    country = row['country']
    admin1 = row['admin1']
    cruiseship = ''
    
    cname = name_dict.get(country)
    if cname != None:
        if cname[1] == 'Country':
            country = cname[0]
        elif cname[1] == 'CruiseShip':
            country = ''
            admin1 = ''
            cruiseship = cname[0]

    aname = name_dict.get(admin1)    
    if aname != None:
        if aname[1] == 'Admin1':
            admin1 = aname[0]
        elif aname[1] == 'Country':
            # dependent territories according to ISO are represented at the country level
            admin1 = ''
            country = aname[0]
        elif aname[1] == 'CruiseShip':
            country = ''
            admin1 = ''
            cruiseship = aname[0]

            
    return country, admin1, cruiseship

df[['country','admin1','cruiseship']] = df.apply(standardize_locations, axis=1, result_type="expand")

Check that country names match GeoNames

countries = pd.read_csv(NEO4J_HOME / "import/00e-GeoNamesCountry.csv")
loc0 = df[['country']].copy()
loc0 = loc0.merge(countries, left_on='country', right_on='name', how='left')

loc0.fillna('', inplace=True)
loc0 = loc0.query("name == ''")
country_mismatches = loc0['country'].unique()
print("Country name mismatches:")
print(country_mismatches)
# TODO comma in string 'Bonaire, Saint Eustatius and Saba' is not handled properly

Check that admin1 names match GeoNames

admin1 = pd.read_csv(NEO4J_HOME / "import/00f-GeoNamesAdmin1.csv")
loc1 = df[['admin1']].copy()
loc1 = loc1.query("admin1 != ''")
loc1 = loc1.merge(admin1, left_on='admin1', right_on='name', how='left')

loc1.fillna('', inplace=True)
loc1 = loc1.query("name == ''")
loc1_mismatches = loc1['admin1'].unique()
print("Admin1 name mismatches:")
print(loc1_mismatches)

#### Save cases for countries

cases_country = df[(df['country'] != '') & (df['admin1'] == '')]
cases_country = cases_country[['country', 'date', 'cummulativeConfirmed', 'cummulativeDeaths']]
cases_country.to_csv(NEO4J_HOME / "import/02a-JHUCasesGlobalCountry.csv", index=False)
cases_country.head()

#### Save cases for admin1 divisions

cases_admin1 = df[(df['country'] != '') & (df['admin1'] != '')]
cases_admin1 = cases_admin1[['country', 'admin1', 'date', 'cummulativeConfirmed', 'cummulativeDeaths']]
cases_admin1.to_csv(NEO4J_HOME / "import/02a-JHUCasesGlobalAdmin1.csv", index=False)
cases_admin1.head()

#### Save cases for cruise ships

cases_cruiseship = df[df['cruiseship'] != '']
cases_cruiseship = cases_cruiseship[['cruiseship', 'date', 'cummulativeConfirmed', 'cummulativeDeaths']]
cases_cruiseship.to_csv(NEO4J_HOME / "import/02a-JHUCasesGlobalCruiseShip.csv", index=False)
cases_cruiseship.head()
