==> notebooks/3-ExampleQueriesBioentities.ipynb <==
# Example queries for biological entities on COVID-19 Knowledge Graph
[Work in progress]

This notebook demonstrates how to run Cypher queries in a Jupyter Notebook by connecting to a database server.

import os
import time
import pandas as pd
from py2neo import Graph

pd.options.display.max_rows = None  # display all rows
pd.options.display.max_columns = None  # display all columsns

graph = Graph("bolt://132.249.238.185:7687", user="reader", password="demo")

### List Node Metadata

query = """
MATCH (n:NodeMetadata)
RETURN n.name, n.shortDescription, n.description, n.example, n.details
"""
graph.run(query).to_data_frame()

### List Organisms in KG

query = """
MATCH (p:Organism)
RETURN p.name as name, p.scientificName as scientificName, p.id as taxonomy
"""
graph.run(query).to_data_frame()

### List Coronavirus Outbreaks

query = """
MATCH (p:Organism)-[:CAUSES]->(o:Outbreak)
RETURN p.name as name, p.scientificName as scientificName, p.id as taxonomy, o.id as outbreak, o.startDate as startDate
"""
graph.run(query).to_data_frame()

### List Strains that are mentioned in PubMed Central Articles

query = """
MATCH (p:Publication)-[:MENTIONS]->(s:Strain)<-[:CARRIES]-(h:Host)
RETURN p.id as pmc, s.name as name, s.collectionDate  as collectionDate, h.name as host, h.id as host_id
ORDER by s.collectionDate
"""
graph.run(query).to_data_frame().head(20)

### List Gene and Protein information for Reference Genome
This query lists the genes and proteins encoded by the SARS-CoV-2 reference genome. This is the first genome of SARS-CoV-2 collected in Wuhan on Dec. 5, 2019.

query = """
MATCH (s:Strain)-[:HAS]->(g:Gene)-[:ENCODES]->(p:Protein)
RETURN s.id as referenceGenome, s.name as name, s.collectionDate  as collectionDate, 
       g.name as gene, g.id as geneId, p.name as protein, p.id as protein_id 
ORDER by s.collectionDate
"""
graph.run(query).to_data_frame()

### Find Mutations in the SARS-CoV-2 "S" gene
The S gene is translated by the human or other species to the spike protein, which binds the human ACE2 receptor to gain entrance into the cell. Here we compare the mutations (missense mutations) found at different locations.

query = """
MATCH (g:Gene{name:'S'})-[:HAS_VARIANT]->(v:Variant{variantConsequence:'missense_variant'})
<-[:HAS_VARIANT]-(s:Strain)-[:FOUND_IN]->(l:Location)-[:IN*]->(r:USRegion{name:$region}) 
RETURN v.name AS mutation, l.name AS location, r.name AS region
ORDER BY v.name, l.name
"""

##### Variants in the US West Region

region = 'West Region'
graph.run(query, region=region).to_data_frame().head(100)

##### Variants in the US Northeast Region

region = 'Northeast Region'
graph.run(query, region=region).to_data_frame().head(100)

### Cases in a specific County (Admin2)

admin2 = 'San Diego County'

query = """
MATCH (c:Cases{date: date("2020-06-04")})-[:REPORTED_IN]->(a:Admin2{name: $admin2})
RETURN a.name as name, c.cummulativeConfirmed as confirmed, c.cummulativeDeaths as deaths
"""
graph.run(query, admin2=admin2).to_data_frame()

### Aggregate cases by State (Admin1)

query = """
MATCH (o:Outbreak{id: "COVID-19"})<-[:RELATED_TO]-(c:Cases{date: date("2020-06-04")})-[:REPORTED_IN]->(a:Admin2)-[:IN]->(a1:Admin1)
RETURN a1.name as state, sum(c.cummulativeDeaths) as deaths, sum(c.cummulativeConfirmed) as confirmed
ORDER BY deaths DESC
"""
graph.run(query).to_data_frame()
